Move IndexCreatingLoadStrategyFactory out of SearchesTest
