+++ graylog2-server/src/test/java/org/graylog2/indexer/searches/SearchesTest.java
@@ -19,26 +19,17 @@ package org.graylog2.indexer.searches;
 import com.codahale.metrics.Histogram;
 import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.Timer;
-import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.ImmutableSortedSet;
 import com.lordofthejars.nosqlunit.annotation.UsingDataSet;
-import com.lordofthejars.nosqlunit.core.DatabaseOperation;
 import com.lordofthejars.nosqlunit.core.LoadStrategyEnum;
-import com.lordofthejars.nosqlunit.core.LoadStrategyFactory;
-import com.lordofthejars.nosqlunit.core.LoadStrategyOperation;
-import com.lordofthejars.nosqlunit.core.ReflectionLoadStrategyFactory;
 import com.lordofthejars.nosqlunit.elasticsearch.ElasticsearchRule;
 import com.lordofthejars.nosqlunit.elasticsearch.EmbeddedElasticsearch;
-import org.elasticsearch.action.admin.indices.exists.indices.IndicesExistsResponse;
 import org.elasticsearch.client.Client;
-import org.elasticsearch.client.IndicesAdminClient;
 import org.elasticsearch.indices.IndexMissingException;
 import org.elasticsearch.search.SearchHit;
 import org.graylog2.Configuration;
-import org.graylog2.configuration.ElasticsearchConfiguration;
 import org.graylog2.indexer.Deflector;
-import org.graylog2.indexer.IndexMapping;
-import org.graylog2.indexer.indices.Indices;
+import org.graylog2.indexer.nosqlunit.IndexCreatingLoadStrategyFactory;
 import org.graylog2.indexer.ranges.IndexRange;
 import org.graylog2.indexer.ranges.IndexRangeComparator;
 import org.graylog2.indexer.ranges.IndexRangeService;
@@ -48,7 +39,6 @@ import org.graylog2.indexer.results.HistogramResult;
 import org.graylog2.indexer.results.TermsResult;
 import org.graylog2.indexer.results.TermsStatsResult;
 import org.graylog2.indexer.searches.timeranges.AbsoluteRange;
-import org.graylog2.plugin.database.validators.Validator;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
 import org.junit.Before;
@@ -439,69 +429,4 @@ public class SearchesTest {
         searches.timestampStatsOfIndex("does-not-exist");
     }
 
-    public static class IndexCreatingLoadStrategyFactory implements LoadStrategyFactory {
-        private final LoadStrategyFactory loadStrategyFactory;
-        private final Set<String> indexNames;
-
-        public IndexCreatingLoadStrategyFactory(Set<String> indexNames) {
-            this.loadStrategyFactory = new ReflectionLoadStrategyFactory();
-            this.indexNames = ImmutableSet.copyOf(indexNames);
-        }
-
-        @Override
-        public LoadStrategyOperation getLoadStrategyInstance(LoadStrategyEnum loadStrategyEnum, DatabaseOperation databaseOperation) {
-            return loadStrategyFactory.getLoadStrategyInstance(
-                    loadStrategyEnum,
-                    new IndexCreatingDatabaseOperation(databaseOperation, indexNames));
-        }
-    }
-
-    public static class IndexCreatingDatabaseOperation implements DatabaseOperation<Client> {
-        private final DatabaseOperation<Client> databaseOperation;
-        private final Client client;
-        private final Set<String> indexes;
-
-        public IndexCreatingDatabaseOperation(DatabaseOperation<Client> databaseOperation, Set<String> indexes) {
-            this.databaseOperation = databaseOperation;
-            this.client = databaseOperation.connectionManager();
-            this.indexes = ImmutableSet.copyOf(indexes);
-        }
-
-        @Override
-        public void insert(InputStream dataScript) {
-            final IndicesAdminClient indicesAdminClient = client.admin().indices();
-            final String[] indexNames = indexes.toArray(new String[indexes.size()]);
-            IndicesExistsResponse indicesExistsResponse = indicesAdminClient.prepareExists(indexNames)
-                    .execute()
-                    .actionGet();
-
-            if (indicesExistsResponse.isExists()) {
-                client.admin().indices().prepareDelete(indexNames).execute().actionGet();
-            }
-
-            Indices indices = new Indices(client, new ElasticsearchConfiguration(), new IndexMapping(client));
-            for (String index : indexes) {
-                if (!indices.create(index)) {
-                    throw new IllegalStateException("Couldn't create index " + index);
-                }
-            }
-
-            databaseOperation.insert(dataScript);
-        }
-
-        @Override
-        public void deleteAll() {
-            databaseOperation.deleteAll();
-        }
-
-        @Override
-        public boolean databaseIs(InputStream expectedData) {
-            return databaseOperation.databaseIs(expectedData);
-        }
-
-        @Override
-        public Client connectionManager() {
-            return databaseOperation.connectionManager();
-        }
-    }
 }
